requirements.txt
langgraph>=0.0.25
langchain-core>=0.1.15
langchain-google-genai>=0.0.5
beautifulsoup4>=4.12.2
requests>=2.31.0
typing-extensions>=4.8.0
python-dotenv>=1.0.0
google-generativeai>=0.4.0 
crawl4ai>=0.4.247 
fastapi>=0.95.0
uvicorn[standard]>=0.22.0 